CHORE: redisConfig 파일 수정

+++ Backend/alcol/rank-service/src/main/java/com/alcol/rankservice/config/RedisConfig.java
@@ -35,7 +35,7 @@ public class RedisConfig {
 //        redisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer());
 
         redisTemplate.setHashKeySerializer(new StringRedisSerializer());
-        redisTemplate.setHashValueSerializer(new GenericToStringSerializer<>(Integer.class));
+        redisTemplate.setHashValueSerializer(new GenericToStringSerializer<>(Long.class));
         return redisTemplate;
     }
 }
